// src/types/protocol.ts
/**
 * SMTP Status Codes as per RFC 5321 and RFC 3463
 */
export enum SMTPCode {
  SYSTEM_STATUS = 211,
  HELP_MESSAGE = 214,
  SERVICE_READY = 220,
  SERVICE_CLOSING = 221,
  AUTH_SUCCESS = 235,
  ACTION_COMPLETED = 250,
  USER_NOT_LOCAL_WILL_FORWARD = 251,
  START_INPUT = 354,
  SERVICE_NOT_AVAILABLE = 421,
  ACTION_NOT_TAKEN = 450,
  ACTION_ABORTED_LOCAL_ERROR = 451,
  INSUFFICIENT_STORAGE = 452,
  UNRECOGNIZED_COMMAND = 500,
  SYNTAX_ERROR_PARAMETERS = 501,
  COMMAND_NOT_IMPLEMENTED = 502,
  BAD_SEQUENCE = 503,
  PARAMETER_NOT_IMPLEMENTED = 504,
  AUTHENTICATION_REQUIRED = 530,
  AUTH_CREDENTIALS_INVALID = 535,
  MAILBOX_UNAVAILABLE = 550,
  USER_NOT_LOCAL = 551,
  EXCEEDED_STORAGE = 552,
  MAILBOX_NAME_NOT_ALLOWED = 553,
  TRANSACTION_FAILED = 554,
}

/**
 * Standard SMTP Commands
 */
export type SMTPCommand = 'HELO' | 'EHLO' | 'AUTH' | 'MAIL' | 'RCPT' | 'DATA' | 'QUIT' | 'RSET' | 'NOOP' | 'STARTTLS';

/**
 * SMTPSession Finite State Machine Enumeration
 */
export enum SessionState {
  CONNECTED,
  AUTHENTICATING,
  AUTHENTICATED,
  MAIL_FROM,
  RCPT_TO,
  DATA_WRITING,
  QUIT,
}